fix responsive, #319
